devel-requirements.txt
pytest-django
flake8
mypy
python-language-server

